requirements.txt
python-telegram-bot
requests
openpyxl
redis
